Move repository to correct package.
